% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvlm_excampledata.R
\name{bvlm_excampledata}
\alias{bvlm_excampledata}
\title{Function for generating a data excample used in bvlm.}
\usage{
bvlm_excampledata(seed, rhoErr)
}
\arguments{
\item{seed:}{Used to control "set.seed".}

\item{rhoErr:}{korrelation in error covariance matrix}
}
\value{
A list bvlm_excample with three elements:
}
\description{
Function for generating a data excample used in bvlm.
}
\section{Fields}{

\describe{
\item{\code{data:}}{A data frame containing the elements Y(100 x 2), with 85 missing
x1 and x2 (both numeric vectors of length 100) and f1 a "factor vector" with
3 levels of length 100}

\item{\code{data_test:}}{Like data, but lacking Y.}

\item{\code{Parameters:}}{A list with 3 elements of "true parameters".
Beta (5 x 2), Sigma and Lambda  (both(2 x 2))}
}}

\references{
Gangsei, Almøy and Sæbø (2019). Linear Regression
 with Bivariate Response Variable Containing Missing Data.
 Strategies to Increase Prediction Precision.
}
\author{
Lars Erik Gangsei, Trygve Almøy and Solve Sæbø
}
