% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evopt.R
\name{evopt}
\alias{evopt}
\title{Function for maximizing model evidxence (Empirical Bayes) in linear
regression model with bivariate response where one of the responses
contains missing data. Used as input in non linear optimizer "nlm"}
\usage{
evopt(aagg, QQ, Beta, XtX, nn, ag_max = c(Inf, Inf))
}
\arguments{
\item{aagg}{A vector of length 2 containing the "alpha" and "gamma hyperpar}

\item{QQ}{A matrix (2 x 2) or scalar}

\item{Beta}{Betaparameter}

\item{XtX}{Gram matrix}

\item{nn}{Number of observations.}

\item{ag_max}{Prior set max values for alpha and gamma.}
}
\value{
the negative logarithm of the Model evidence avaluated at with
hyperparameters alpha and gamma given in aagg
}
\description{
Function for maximizing model evidxence (Empirical Bayes) in linear
regression model with bivariate response where one of the responses
contains missing data. Used as input in non linear optimizer "nlm"
}
\references{
Gangsei, Almøy and Sæbø (2019). Linear Regression
 with Bivariate Response Variable Containing Missing Data.
 Strategies to Increase Prediction Precision.
}
\author{
Lars Erik Gangsei, Trygve Almøy and Solve Sæbø
}
