% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic_calc.R
\name{fic_calc}
\alias{fic_calc}
\title{Function returning FIC values for wide and narrow model using ML
estimates in the bivariate regression model exept for lambda12, which
is an argument to the function. Used for making figures.}
\usage{
fic_calc(lambda12, xxN, Beta11h, Beta21h, XtX1_inv, XtX2, QQ, qq3, qq5,
  qq6, nn1, nn2)
}
\arguments{
\item{QQ, }{qq3,qq5,qq6: Quadratic terms, see manuscript for details}

\item{nn1, nn2:}{Total number of observations and number of full observations}

\item{lambda12:}{The value for lambda12 for which to evaluate the FIC values.}

\item{xxN:}{a vector (p x 1) for which the forcus parameter is defined as
xxN^t x Beta2. Note that Beta2 is not input in the function.}

\item{Beta11h:}{OLS estimates for Beta1 based on n1 (all) observations}

\item{Beta21h:}{OLS estimates for Beta1 based on n2 (full/ not missing data)
observations.}

\item{XtX2:}{Gram matrix for predictors based on full/ not missing data}

\item{XtX1_inv:}{Inverse gram matrix for predictors based on all data}
}
\value{
A vector of length two containing two elements, the FIC vales for the
}
\description{
Function returning FIC values for wide and narrow model using ML
estimates in the bivariate regression model exept for lambda12, which
is an argument to the function. Used for making figures.
}
\references{
Gangsei, Almøy and Sæbø (2019). Linear Regression
 with Bivariate Response Variable Containing Missing Data.
 Strategies to Increase Prediction Precision.
}
\author{
Lars Erik Gangsei, Trygve Almøy and Solve Sæbø
}
