% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello from BMLM}
\usage{
hello(name)
}
\arguments{
\item{name}{Name of the person you want to Halloo}
}
\value{
Print hello message to the screen
}
\description{
Hello from BMLM
}
\examples{
hello("user")
}
