% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda12_fit.R
\name{lambda12_fit}
\alias{lambda12_fit}
\title{Function fitting the value for lambda12 so that FIC value for the
 input focus parameter xxN^t x Beta2 is maximized}
\usage{
lambda12_fit(xxN, Beta11h, Beta21h, XtX1_inv, XtX2, QQ, qq3, qq5, nn1, nn2)
}
\arguments{
\item{QQ, qq3, qq5:}{Quadratic terms, see manuscript for details}

\item{nn1, nn2:}{Total number of observations and number of full observations}

\item{xxN:}{a vector (p x 1) for which the forcus parameter is defined as
xxN^t x Beta2. Note that Beta2 is not input in the function.}

\item{Beta11h:}{OLS estimates for Beta1 based on n1 (all) observations}

\item{Beta21h:}{OLS estimates for Beta1 based on n2 (full/ not missing data)
observations.}

\item{XtX2:}{Gram matrix for predictors based on full/ not missing data}

\item{XtX1_inv:}{Inverse gram matrix for predictors based on all data}
}
\value{
two named vectors:
}
\description{
Function fitting the value for lambda12 so that FIC value for the
 input focus parameter xxN^t x Beta2 is maximized
}
\section{Fields}{

\describe{
\item{\code{lambda12:}}{the lambda12 estimate for the Narrow, FIC and ML situation respectively}

\item{\code{fic:}}{The fic values assosiated with the Narrow, FIC and ML situation.
FIC value for narrow is calculated besed on comparisoin with lambda12FIC}
}}

\references{
Gangsei, Almøy and Sæbø (2019). Linear Regression
 with Bivariate Response Variable Containing Missing Data.
 Strategies to Increase Prediction Precision.
}
\author{
Lars Erik Gangsei, Trygve Almøy and Solve Sæbø
}
