% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbvlm.R
\name{pbvlm}
\alias{pbvlm}
\title{Function returning the probability density value (pdf-value) for bivariate linear regression at parameter value theta}
\usage{
pbvlm(theta, XtX1, XtX2, Betah1, Betah2, QQ, qq3, nn1, nn2, log_r = TRUE)
}
\arguments{
\item{theta:}{A vector of length 2p + 3 containing values for lambda11,}

\item{XtX1:}{Gram matrix (p x p) based on X1.}

\item{XtX2:}{Gram matrix (p x p) based on X2.}

\item{Betah1:}{(p x 1) ols estimate for Beta1}

\item{Betah2:}{(px 2) ols estimate for Beta1 and Beta2 based on n2}

\item{QQ:}{2 x 2 squared residual matrix}

\item{qq3:}{SSE for linear model based on XX1}

\item{nn1:}{Number of observations (X1).}

\item{nn2:}{Number of observations (X2).}

\item{log_r:}{Logical. When TRUE (default) log likelihood is returned.}
}
\value{
The likelihood (log likelihood is default) for model
and data in question avaluated at parameterestimates defined in theta
}
\description{
Function returning the probability density value (pdf-value) for bivariate linear regression at parameter value theta
}
\references{
Gangsei, Almøy and Sæbø (2019). Linear Regression
 with Bivariate Response Variable Containing Missing Data.
 Strategies to Increase Prediction Precision.
}
\author{
Lars Erik Gangsei, Trygve Almøy and Solve Sæbø
}
