% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bvlm.R
\name{predict_bvlm}
\alias{predict_bvlm}
\title{Function returning predicted values for bivariate linear regression}
\usage{
predict_bvlm(Obj, newdata = NULL, ficpredict = FALSE)
}
\arguments{
\item{Obj:}{An object of class 'bvlm'}

\item{newdata:}{A data frame containing observations that shall be predicted.
Names corresponding to the formula of Obj.}

\item{ficpredict:}{Logical value indicating if fic prediction shall be applied.
Default is FALSE. Only possible if Obj$method = 'ML'}
}
\value{
A named list
}
\description{
Function returning predicted values for bivariate linear regression
}
\section{Fields}{

\describe{
\item{\code{fitted:}}{a n x 2 matrix with fitted values for both responses}

\item{\code{lambda12:}}{Only if ficpredict is TRUE. A n x 3 matrix fic
values for each new prediction assosiated with  Narrow, FIC and ML situation.}

\item{\code{fic:}}{Only if ficpredict is TRUE. A n x 3 matrix containing
fic values for each new prediction assosiated with  Narrow, FIC
and ML situation.}

\item{\code{ficfitted2:}}{Only if ficpredict is TRUE. A n x 3 matrix containing
predicted values for the second response assosiated with the  Narrow, FIC
and ML situation.}
}}

\references{
Gangsei, Almøy and Sæbø (2019). Linear Regression
 with Bivariate Response Variable Containing Missing Data.
 Strategies to Increase Prediction Precision.
}
\author{
Lars Erik Gangsei, Trygve Almøy and Solve Sæbø
}
